Add Stores API reading (stores and stores with seasons)

+++ backend/src/main/kotlin/adapters/StoresGateway.kt
@@ -5,6 +5,6 @@ import entities.Store
 
 interface StoresGateway {
     fun getStores(): List<Store>
-    fun getStoresAndSeasons(): Map<Int, Season>
+    fun getStoresAndSeasons(): Map<Long, Season>
     fun getCSV()
 }
